Advanced Querying Complete
Updated the query system so that more advanced dynamic queries can be
made.
- Allows for domain comparison groups, as described by the user.
- Enhanced the system to work with POST requests. Note that to do advanced
  queries, the request must be POST.
- Has been tested internally.
- Fixed display of data on slo_covered_by_kts view.

This should wrap up the backend for the query system.
NOTE: Dynamic queries that involve any tables that are not directly
related to the Student table will not work. Due to the design of the
database, these queries cannot be handled dynamically!